badminton_club/settings/production.py
from .base import *  # noqa
import environ

env = environ.Env()

DEBUG = False

SECURE_SSL_REDIRECT = True
SECURE_HSTS_SECONDS = 31536000
SECURE_HSTS_INCLUDE_SUBDOMAINS = True
SECURE_HSTS_PRELOAD = True
SESSION_COOKIE_SECURE = True
CSRF_COOKIE_SECURE = True
SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')

CONN_MAX_AGE = 600

EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
