requirements.txt
Django>=5.1,<6.0
psycopg2-binary>=2.9
gunicorn>=22.0
django-environ>=0.11
Pillow>=10.0
